# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talent Bridge API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get list of jobs",
                "operationId": "2a1f2306c8b5239d199c9a5bf97c9ced",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query for job title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Create a new job",
                "operationId": "08a4e2ab29ca916a6bb7d999d11d16a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "min_experience_year",
                                    "number_of_employee",
                                    "duration",
                                    "status",
                                    "type_job",
                                    "type_salary",
                                    "system"
                                ],
                                "properties": {
                                    "min_experience_year": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "number_of_employee": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "6 months"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "open"
                                    },
                                    "type_job": {
                                        "type": "string",
                                        "example": "full-time"
                                    },
                                    "type_salary": {
                                        "type": "string",
                                        "example": "monthly"
                                    },
                                    "system": {
                                        "type": "string",
                                        "example": "remote"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Get job details",
                "operationId": "8e49157694025e1aa63a4cb529c41cad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "massage": {
                                            "type": "string",
                                            "example": "job apa sih yang kamu cari"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Job"
                ],
                "summary": "Update a job",
                "operationId": "7317872995b9bc5a2c06a3e7aef93ecd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "min_experience_year",
                                    "number_of_employee",
                                    "duration",
                                    "status",
                                    "type_job",
                                    "type_salary",
                                    "system"
                                ],
                                "properties": {
                                    "min_experience_year": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "number_of_employee": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "12 months"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "closed"
                                    },
                                    "type_job": {
                                        "type": "string",
                                        "example": "part-time"
                                    },
                                    "type_salary": {
                                        "type": "string",
                                        "example": "hourly"
                                    },
                                    "system": {
                                        "type": "string",
                                        "example": "on-site"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Job"
                ],
                "summary": "Delete a job",
                "operationId": "6ffd6ae6a4d42ac3e97b16b55af6b7a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "massage": {
                                            "type": "string",
                                            "example": "deleted succesfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "massage": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "min_experience_year": {
                        "type": "integer",
                        "example": 2
                    },
                    "number_of_employee": {
                        "type": "integer",
                        "example": 5
                    },
                    "duration": {
                        "type": "string",
                        "example": "6 months"
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    },
                    "type_job": {
                        "type": "string",
                        "example": "full-time"
                    },
                    "type_salary": {
                        "type": "string",
                        "example": "monthly"
                    },
                    "system": {
                        "type": "string",
                        "example": "remote"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}